Fix warning: assuming signed overflow does not occur when simplifying conditional to constant [-Wstrict-overflow]

+++ modules/mbt/src/klt/vpMbKltTracker.cpp
@@ -416,7 +416,7 @@ vpMbKltTracker::setCameraParameters(const vpCameraParameters& camera)
 void
 vpMbKltTracker::setPose(const vpImage<unsigned char> &I, const vpHomogeneousMatrix& cdMo)
 {
-  if(kltCylinders.size() != 0)
+  if((int)(kltCylinders.size()) != 0)
   {
     std::cout << "WARNING: Cannot set pose when model contains cylinder(s). This feature is not implemented yet." << std::endl;
     std::cout << "Tracker will be reinitialized with the given pose." << std::endl;
